split registry sFn into two smaller state functions
